+++ cli/azd/pkg/auth/manager.go
@@ -10,21 +10,25 @@ import (
 	"errors"
 	"fmt"
 	"log"
+	"net/url"
 	"os"
 	"path/filepath"
 	"strconv"
+	"strings"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	azcloud "github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/public"
 	"github.com/azure/azure-dev/cli/azd/internal/tracing"
 	"github.com/azure/azure-dev/cli/azd/internal/tracing/fields"
-	"github.com/azure/azure-dev/cli/azd/pkg/azure"
+	"github.com/azure/azure-dev/cli/azd/pkg/cloud"
 	"github.com/azure/azure-dev/cli/azd/pkg/config"
 	"github.com/azure/azure-dev/cli/azd/pkg/github"
 	"github.com/azure/azure-dev/cli/azd/pkg/httputil"
 	"github.com/azure/azure-dev/cli/azd/pkg/input"
+	"github.com/azure/azure-dev/cli/azd/pkg/oneauth"
 	"github.com/azure/azure-dev/cli/azd/pkg/osutil"
 	"github.com/azure/azure-dev/cli/azd/pkg/output"
 	"github.com/azure/azure-dev/cli/azd/pkg/output/ux"
@@ -48,22 +52,10 @@ const cUseAzCliAuthKey = "auth.useAzCliAuth"
 // auth related configuration information (e.g. the home account id of the current user). This information is not secret.
 const cAuthConfigFileName = "auth.json"
 
-// cDefaultAuthority is the default authority to use when a specific tenant is not presented. We use "organizations" to
-// allow both work/school accounts and personal accounts (this matches the default authority the `az` CLI uses when logging
-// in).
-const cDefaultAuthority = "https://login.microsoftonline.com/organizations"
-
 const cUseCloudShellAuthEnvVar = "AZD_IN_CLOUDSHELL"
 
 const cExternalAuthEndpointEnvVarName = "AZD_AUTH_ENDPOINT"
 const cExternalAuthKeyEnvVarName = "AZD_AUTH_KEY"
-
-// The scopes to request when acquiring our token during the login flow or when requesting a token to validate if the client
-// is logged in.
-var LoginScopes = []string{azure.ManagementScope}
-var loginScopesMap = map[string]struct{}{
-	azure.ManagementScope: {},
-}
 
 // HttpClient interface as required by MSAL library.
 type HttpClient interface {
@@ -91,6 +83,7 @@ type HttpClient interface {
 type Manager struct {
 	publicClient        publicClient
 	publicClientOptions []public.Option
+	cloud               *cloud.Cloud
 	configManager       config.FileConfigManager
 	userConfigManager   config.UserConfigManager
 	credentialCache     Cache
@@ -102,6 +95,7 @@ type Manager struct {
 func NewManager(
 	configManager config.FileConfigManager,
 	userConfigManager config.UserConfigManager,
+	cloud *cloud.Cloud,
 	httpClient HttpClient,
 	console input.Console,
 ) (*Manager, error) {
@@ -120,9 +114,14 @@ func NewManager(
 		return nil, fmt.Errorf("creating msal cache root: %w", err)
 	}
 
+	authorityUrl, err := url.JoinPath(cloud.Configuration.ActiveDirectoryAuthorityHost, "organizations")
+	if err != nil {
+		return nil, fmt.Errorf("joining authority url: %w", err)
+	}
+
 	options := []public.Option{
 		public.WithCache(newCache(cacheRoot)),
-		public.WithAuthority(cDefaultAuthority),
+		public.WithAuthority(authorityUrl),
 		public.WithHTTPClient(httpClient),
 	}
 
@@ -136,6 +135,7 @@ func NewManager(
 	return &Manager{
 		publicClient:        &msalPublicClientAdapter{client: &publicClientApp},
 		publicClientOptions: options,
+		cloud:               cloud,
 		configManager:       configManager,
 		userConfigManager:   userConfigManager,
 		credentialCache:     newCredentialCache(authRoot),
@@ -145,11 +145,32 @@ func NewManager(
 	}, nil
 }
 
+func LoginScopes(cloud *cloud.Cloud) []string {
+	resourceManagerUrl := cloud.Configuration.Services[azcloud.ResourceManager].Endpoint
+	return []string{
+		fmt.Sprintf("%s//.default", resourceManagerUrl),
+	}
+}
+
+func (m *Manager) LoginScopes() []string {
+	return LoginScopes(m.cloud)
+}
+
+func loginScopesMap(cloud *cloud.Cloud) map[string]struct{} {
+	resourceManagerUrl := cloud.Configuration.Services[azcloud.ResourceManager].Endpoint
+
+	return map[string]struct{}{resourceManagerUrl: {}}
+}
+
 // EnsureLoggedInCredential uses the credential's GetToken method to ensure an access token can be fetched.
 // On success, the token we fetched is returned.
-func EnsureLoggedInCredential(ctx context.Context, credential azcore.TokenCredential) (*azcore.AccessToken, error) {
+func EnsureLoggedInCredential(
+	ctx context.Context,
+	credential azcore.TokenCredential,
+	cloud *cloud.Cloud,
+) (*azcore.AccessToken, error) {
 	token, err := credential.GetToken(ctx, policy.TokenRequestOptions{
-		Scopes: LoginScopes,
+		Scopes: LoginScopes(cloud),
 	})
 	if err != nil {
 		return &azcore.AccessToken{}, err
@@ -213,6 +234,23 @@ func (m *Manager) CredentialForCurrentUser(
 	}
 
 	if currentUser.HomeAccountID != nil {
+		if currentUser.FromOneAuth {
+			tenant := options.TenantID
+			if tenant == "" {
+				tenant = "organizations"
+			}
+
+			authority, err := url.JoinPath(m.cloud.Configuration.ActiveDirectoryAuthorityHost, tenant)
+			if err != nil {
+				return nil, fmt.Errorf("joining authority url: %w", err)
+			}
+
+			return oneauth.NewCredential(authority, cAZD_CLIENT_ID, oneauth.CredentialOptions{
+				HomeAccountID: *currentUser.HomeAccountID,
+				NoPrompt:      options.NoPrompt,
+			})
+		}
+
 		accounts, err := m.publicClient.Accounts(ctx)
 		if err != nil {
 			return nil, err
@@ -220,9 +258,9 @@ func (m *Manager) CredentialForCurrentUser(
 		for i, account := range accounts {
 			if account.HomeAccountID == *currentUser.HomeAccountID {
 				if options.TenantID == "" {
-					return newAzdCredential(m.publicClient, &accounts[i]), nil
+					return newAzdCredential(m.publicClient, &accounts[i], m.cloud), nil
 				} else {
-					newAuthority := "https://login.microsoftonline.com/" + options.TenantID
+					newAuthority := m.cloud.Configuration.ActiveDirectoryAuthorityHost + options.TenantID
 
 					newOptions := make([]public.Option, 0, len(m.publicClientOptions)+1)
 					newOptions = append(newOptions, m.publicClientOptions...)
@@ -236,7 +274,7 @@ func (m *Manager) CredentialForCurrentUser(
 						return nil, err
 					}
 
-					return newAzdCredential(&msalPublicClientAdapter{client: &clientWithNewTenant}, &accounts[i]), nil
+					return newAzdCredential(&msalPublicClientAdapter{client: &clientWithNewTenant}, &accounts[i], m.cloud), nil
 				}
 			}
 		}
@@ -330,7 +368,7 @@ func (m *Manager) GetLoggedInServicePrincipalTenantID(ctx context.Context) (*str
 				return nil, err
 			}
 
-			token, err := EnsureLoggedInCredential(ctx, credential)
+			token, err := EnsureLoggedInCredential(ctx, credential, m.cloud)
 			if err != nil {
 				return nil, err
 			}
@@ -361,10 +399,14 @@ func (m *Manager) GetLoggedInServicePrincipalTenantID(ctx context.Context) (*str
 func (m *Manager) newCredentialFromClientSecret(
 	tenantID string,
 	clientID string,
-	clientSecret string) (azcore.TokenCredential, error) {
+	clientSecret string,
+) (azcore.TokenCredential, error) {
 	options := &azidentity.ClientSecretCredentialOptions{
 		ClientOptions: policy.ClientOptions{
 			Transport: m.httpClient,
+			// TODO: Inject client options instead? this can be done if we're OK
+			// using the default user agent string.
+			Cloud: m.cloud.Configuration,
 		},
 	}
 	cred, err := azidentity.NewClientSecretCredential(tenantID, clientID, clientSecret, options)
@@ -393,6 +435,9 @@ func (m *Manager) newCredentialFromClientCertificate(
 	options := &azidentity.ClientCertificateCredentialOptions{
 		ClientOptions: policy.ClientOptions{
 			Transport: m.httpClient,
+			// TODO: Inject client options instead? this can be done if we're OK
+			// using the default user agent string.
+			Cloud: m.cloud.Configuration,
 		},
 	}
 	cred, err := azidentity.NewClientCertificateCredential(
@@ -416,6 +461,9 @@ func (m *Manager) newCredentialFromFederatedTokenProvider(
 	options := &azidentity.ClientAssertionCredentialOptions{
 		ClientOptions: policy.ClientOptions{
 			Transport: m.httpClient,
+			// TODO: Inject client options instead? this can be done if we're OK
+			// using the default user agent string.
+			Cloud: m.cloud.Configuration,
 		},
 	}
 	cred, err := azidentity.NewClientAssertionCredential(
@@ -457,7 +505,7 @@ func (m *Manager) LoginInteractive(
 	scopes []string,
 	options *LoginInteractiveOptions) (azcore.TokenCredential, error) {
 	if scopes == nil {
-		scopes = LoginScopes
+		scopes = m.LoginScopes()
 	}
 	acquireTokenOptions := []public.AcquireInteractiveOption{}
 	if options == nil {
@@ -486,13 +534,29 @@ func (m *Manager) LoginInteractive(
 		return nil, err
 	}
 
-	return newAzdCredential(m.publicClient, &res.Account), nil
+	return newAzdCredential(m.publicClient, &res.Account, m.cloud), nil
+}
+
+func (m *Manager) LoginWithOneAuth(ctx context.Context, tenantID string, scopes []string) error {
+	if len(scopes) == 0 {
+		scopes = m.LoginScopes()
+	}
+
+	authority := m.cloud.Configuration.ActiveDirectoryAuthorityHost + tenantID
+	accountID, err := oneauth.LogIn(authority, cAZD_CLIENT_ID, strings.Join(scopes, " "))
+	if err == nil {
+		err = m.saveUserProperties(&userProperties{
+			FromOneAuth:   true,
+			HomeAccountID: &accountID,
+		})
+	}
+	return err
 }
 
 func (m *Manager) LoginWithDeviceCode(
 	ctx context.Context, tenantID string, scopes []string, withOpenUrl WithOpenUrl) (azcore.TokenCredential, error) {
 	if scopes == nil {
-		scopes = LoginScopes
+		scopes = m.LoginScopes()
 	}
 	options := []public.AcquireByDeviceCodeOption{}
 	if tenantID != "" {
@@ -542,7 +606,7 @@ func (m *Manager) LoginWithDeviceCode(
 		return nil, err
 	}
 
-	return newAzdCredential(m.publicClient, &res.Account), nil
+	return newAzdCredential(m.publicClient, &res.Account, m.cloud), nil
 
 }
 
@@ -638,13 +702,18 @@ func (m *Manager) Logout(ctx context.Context) error {
 
 	// we are fine to ignore the error here, it just means there's nothing to clean up.
 	currentUser, _ := readUserProperties(cfg)
-
-	// When logged in as a service principal, remove the stored credential
-	if currentUser != nil && currentUser.TenantID != nil && currentUser.ClientID != nil {
-		if err := m.saveLoginForServicePrincipal(
-			*currentUser.TenantID, *currentUser.ClientID, &persistedSecret{},
-		); err != nil {
-			return fmt.Errorf("removing authentication secrets: %w", err)
+	if currentUser != nil {
+		if currentUser.FromOneAuth {
+			if err := oneauth.Logout(cAZD_CLIENT_ID); err != nil {
+				return fmt.Errorf("logging out of OneAuth: %w", err)
+			}
+		} else if currentUser.TenantID != nil && currentUser.ClientID != nil {
+			// When logged in as a service principal, remove the stored credential
+			if err := m.saveLoginForServicePrincipal(
+				*currentUser.TenantID, *currentUser.ClientID, &persistedSecret{},
+			); err != nil {
+				return fmt.Errorf("removing authentication secrets: %w", err)
+			}
 		}
 	}
 
@@ -821,6 +890,8 @@ func (m *Manager) saveSecret(tenantId, clientId string, ps *persistedSecret) err
 }
 
 type CredentialForCurrentUserOptions struct {
+	// NoPrompt controls whether the credential may prompt for user interaction.
+	NoPrompt bool
 	// The tenant ID to use when constructing the credential, instead of the default tenant.
 	TenantID string
 }
@@ -858,6 +929,7 @@ type federatedAuth struct {
 // client).
 type userProperties struct {
 	HomeAccountID *string `json:"homeAccountId,omitempty"`
+	FromOneAuth   bool    `json:"fromOneAuth,omitempty"`
 	ClientID      *string `json:"clientId,omitempty"`
 	TenantID      *string `json:"tenantId,omitempty"`
 }