+++ cli/azd/test/functional/remote_state_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/azure/azure-dev/cli/azd/pkg/auth"
 	"github.com/azure/azure-dev/cli/azd/pkg/azsdk"
 	"github.com/azure/azure-dev/cli/azd/pkg/azsdk/storage"
+	"github.com/azure/azure-dev/cli/azd/pkg/cloud"
 	"github.com/azure/azure-dev/cli/azd/pkg/config"
 	"github.com/azure/azure-dev/cli/azd/pkg/environment"
 	"github.com/azure/azure-dev/cli/azd/test/azdcli"
@@ -74,6 +75,7 @@ func createBlobClient(
 	authManager, err := auth.NewManager(
 		fileConfigManager,
 		config.NewUserConfigManager(fileConfigManager),
+		cloud.AzurePublic(),
 		httpClient, mockContext.Console,
 	)
 	require.NoError(t, err)
@@ -81,11 +83,11 @@ func createBlobClient(
 	credentials, err := authManager.CredentialForCurrentUser(*mockContext.Context, nil)
 	require.NoError(t, err)
 
-	coreClientOptions := azsdk.NewClientOptionsBuilderFactory(httpClient, "azd").
+	coreClientOptions := azsdk.NewClientOptionsBuilderFactory(httpClient, "azd", cloud.AzurePublic()).
 		NewClientOptionsBuilder().
 		BuildCoreClientOptions()
 
-	sdkClient, err := storage.NewBlobSdkClient(credentials, storageConfig, coreClientOptions)
+	sdkClient, err := storage.NewBlobSdkClient(credentials, storageConfig, coreClientOptions, cloud.AzurePublic())
 	require.NoError(t, err)
 	require.NotNil(t, sdkClient)